refactor(portfolio): split component into smaller parts
